Implemented own component for environment card and fixed reactivity issue when changing clusters
